Initial Solution to bigdecimal
This solution performs the sorting properly, but the question requires that the
strings be read in, converted on the fly to bigdecimal, and using the compareto
method to then sort the array and inserting the corresponding string value
based on the comparison. This will preserve the format of the string input.